requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
sentence-transformers==2.7.0
torch==2.2.0
numpy<2.0.0
